Update 1.3
Fix `ArrayIndexOutOfBoundsException` when executing Raw commands with arguments.

##### Added:
New command `/world spawnlogic` - Enables or disables the world spawn point logic.
Customizable restrictions on entity spawns.
Ability to save inventory and experience of the player at death. Read more in the documentation.
##### Added for Forge servers:
Logging the list of mods installed on the client player connects to the server.
Configurable restrictions on allowable mods on the player's client that connects to the server.