Add support NeoForge
The use of Forge classes has been completely removed from the main part of the plugin.  Information about installed mods on the modified server is now collected via mixin.  Forge and NeoForge mods are supported.
Getting a player's mod list is only available on LexForge for now. NeoForge does not have the necessary API. A workaround is needed.