Improved package handling
The reading of packet data has been improved. Non-printable characters are now replaced by spaces after byte to string conversion. If there are extra spaces at the end of the resulting string, they will be removed.
Added the ability to retrieve the name of the client.
Mod and plugin collections are now unmodifiable and labeled `final`.