Add MixinServerWorld
Added the ability to stop ticks in an individual world to the API.  This is similar to the vanilla `/tick freeze' command, which affects all worlds at once.
Through the new interface, you can now get the world's TPS and tick time.  Previously existing methods in the API are marked as deprecated and may be removed in the future.
Renamed mixin classes to reduce the length of their names.